listHistory function now returns an Object
listHistory func in gmailAPI.js changed to now return an object.
This object contains "messageAddedIds" as key, and an array of id's as
value. The ids can then be used to actually get the messages added.

Note that the Google doc says MessageAdded in history list to actually
contain the full message but its not the case. It just provides the id,
threadId & labelIds.