BUg fix in conciseMessage function
Generally, the 'From' header in email is of type:
"Name <xyz@exampleemail.com>"
But in case of some emails with no name is found to have
no angle brackets. It is of type:
"xyz@exampleemail.com"

I was getting one character less in senderEmail after parsing in the later condition,
for example "xyz@exampleemail.co" , 'm' is missing.