Added filter functions, default filter settings
Added filter functions in lib.js file. The filtering based on message
labels is working fine and the changes for that is in index.js file.
The defaultFilter.js file contains the default filter settings which was
used to test the label filtering.

Primary filter functions added are
"filterMessageBasedOnProperty", "filterMessageBasedOnLabels",
"filterMessageBasedOnEmailIds", "filterMessageBasedOnDomains"and
"filterMessage" . Only label filtering and hence
filterMessageBasedOnProperty is tested and working fine. Others should
also work, need to test.