Bug fix in getNewFilteredMessages function
Same data was saved to sync.json and sync-backup.json. The object was
supposed to be edited and present data in sync file should be saved to
backup, while latest history id should be saved to sync.json. But there
was a bug, the object was assigned as reference. Now using the spread
operator to copy the object rather than pointing by reference.